checkout tests from rubybopmaster
